Move createPoints function to common helper
It would be useful for horizon creation
