+++ tests/performance/helpers/request-helpers.js
@@ -42,6 +42,24 @@ export function getRandomIndexInDimension(metadata, annotationDimension) {
   return getRandom(0, getMaxIndexInDimension(metadata, annotationDimension), 1);
 }
 
+/**
+ * Creates set of segmentCount + 1 points that reach from start to end, may be
+ * integers and are relatively equally spaced.
+ */
+export function createEquidistantPoints(start, end, segmentCount, roundToSample = true, sampleStep = 1) {
+  const length = end - start;
+  const segment = length / segmentCount;
+  var points = [];
+  for (let i = 0; i <= segmentCount; i++) {
+    let position = start + i * segment;
+    if (roundToSample) {
+      position = Math.round((position - start) / sampleStep) * sampleStep + start;
+    }
+    points.push(position);
+  }
+  return points;
+}
+
 export function convertDimension(dimension) {
   let converted;
   switch (dimension) {
